chore: release v0.5.5

+++ src/core/clock.rs
@@ -251,7 +251,7 @@ pub fn now_us() -> i64 {
         unsafe {
             libc::clock_gettime(clock_id, &mut ts);
         }
-        (ts.tv_sec as i64) * 1_000_000 + (ts.tv_nsec as i64) / 1_000
+        ts.tv_sec * 1_000_000 + ts.tv_nsec / 1_000
     }
 }
